requirements.txt
streamlit>=1.28.0
crewai>=0.28.0
crewai-tools>=0.1.0
PyYAML>=6.0
pathlib2>=2.3.7
python-dotenv>=1.0.0
